refactor(system): extract SystemBuilder to system/builder.py
Move SystemBuilder out of core.py into its own module. core.py now
contains only the JarvisSystem dataclass, down from 991 to 321 lines.
builder.py is 530 lines of focused construction logic. __init__.py
re-exports SystemBuilder so external 'from openjarvis.system import
SystemBuilder' callers keep working.

Update tests/mcp/test_discovery.py patch target to
openjarvis.system.builder.logger.
