+++ src/openjarvis/system/__init__.py
@@ -1,12 +1,13 @@
 """Top-level system composition: JarvisSystem, SystemBuilder, and helpers."""
 
+from openjarvis.system.builder import SystemBuilder
 from openjarvis.system.bundles import (
     AgentRuntime,
     Observability,
     Scheduling,
     SecurityContext,
 )
-from openjarvis.system.core import JarvisSystem, SystemBuilder
+from openjarvis.system.core import JarvisSystem
 from openjarvis.system.orchestrator import QueryOrchestrator
 from openjarvis.system.protocols import OrchestratorDeps
